+++ idunn/directions/client.py
@@ -1,321 +1,23 @@
-from copy import deepcopy
-from enum import Enum
-from geojson_pydantic.geometries import Geometry, LineString
-from geojson_pydantic.types import Position
-from pydantic.fields import Field
-import requests
 import logging
+import requests
 from datetime import datetime, timedelta
-from starlette.requests import QueryParams
 from fastapi import HTTPException
 from fastapi.responses import JSONResponse
 from pydantic import BaseModel
-from typing import Iterator, List, Optional, Tuple
-from pydantic.class_validators import validator
 from shapely.geometry import Point
-from geojson_pydantic import Feature
-from shapely.ops import unary_union, cascaded_union
-from shapely.geometry import shape
-import shapely.geometry
+from starlette.requests import QueryParams
+from typing import Optional
 
 
 from idunn import settings
+from idunn.datasources.navitia.models import NavitiaResponse
+from idunn.directions.models import DirectionsResponse, IdunnTransportMode
 from idunn.utils.geometry import city_surrounds_polygons
 from idunn.places.base import BasePlace
-from .models import (
-    DirectionsData,
-    DirectionsResponse,
-    DirectionsRoute,
-    RouteLeg,
-    RouteManeuver,
-    RouteStep,
-    RouteSummaryPart,
-    TransportInfo,
-    TransportMode,
-    TransportStop,
-)
 
 logger = logging.getLogger(__name__)
 
 COMBIGO_SUPPORTED_LANGUAGES = {"en", "es", "de", "fr", "it"}
-
-
-class IdunnTransportMode(Enum):
-    CAR = "car"
-    BIKE = "bike"
-    WALKING = "walking"
-    PUBLICTRANSPORT = "publictransport"
-
-    @classmethod
-    def parse(cls, mode: str):
-        if mode in ("driving-traffic", "driving", "car", "car_no_park"):
-            return cls.CAR
-        if mode in ("cycling",):
-            return cls.BIKE
-        if mode in ("walking", "walk"):
-            return cls.WALKING
-        if mode in ("publictransport", "taxi", "vtc", "carpool"):
-            return cls.PUBLICTRANSPORT
-
-        print("UNKNOWN MODE:", mode)
-
-    def to_navitia(self) -> str:
-        if self == self.CAR:
-            return "car_no_park"
-        return self.value
-
-    def to_mapbox(self) -> TransportMode:
-        # NOTE: switch to Python 3.10 for more eleguant matching ?
-        if self == self.CAR:
-            return TransportMode.car
-        if self == self.BIKE:
-            return TransportMode.bicycle
-        if self == self.WALKING:
-            return TransportMode.walk
-        if self == self.PUBLICTRANSPORT:
-            return TransportMode.tram
-
-
-# Mapbox <-> Navitia correspondance
-#  - route <-> journey
-#  - leg <-> section
-#  - step <-> single element of section ? Can be re-cut later ?
-
-
-class Distances(BaseModel):
-    bike: int
-    car: int
-    ridesharing: int
-    taxi: int
-    walking: int
-
-    def overall(self):
-        return self.bike + self.car + self.ridesharing + self.taxi + self.walking
-
-
-class Coordinate(BaseModel):
-    lat: float
-    lon: float
-
-    def to_position(self) -> Position:
-        return (self.lon, self.lat)
-
-
-class Instruction(BaseModel):
-    direction: int
-    duration: int
-    instruction: Optional[str]
-    instruction_start_coordinate: Optional[Coordinate]
-    length: int
-    name: str
-
-    def get_mapbox_modifier(self) -> str:
-        # See https://docs.mapbox.com/api/navigation/directions/#step-maneuver-object
-        mapping = {
-            -135: "sharp left",
-            -90: "left",
-            -45: "slight left",
-            0: "straight",
-            45: "slight right",
-            90: "right",
-            135: "sharp right",
-            180: "uturn",
-        }
-
-        closest = min(
-            *mapping.keys(),
-            key=lambda angle: min(abs(self.direction - angle), abs(360 - self.direction - angle)),
-        )
-
-        return mapping[closest]
-
-
-class DisplayInformations(BaseModel):
-    color: Optional[str]
-    text_color: Optional[str]
-    label: Optional[str]
-
-
-class Section(BaseModel):
-    sec_type: str = Field(..., alias="type")  # TODO: could be an enum
-    mode: Optional[IdunnTransportMode]
-    duration: int
-    path: List[Instruction] = []
-    geojson: Optional[Geometry]  # TODO: always set when not waiting
-    display_informations: Optional[DisplayInformations]
-
-    @validator("mode", pre=True)
-    def mode_from_navitia(cls, v):
-        return IdunnTransportMode.parse(v)
-
-    def cut_linestring(self) -> Iterator[Tuple[Instruction, Geometry]]:
-        """
-        Split path geometry at each instruction and wrap them together
-        """
-        all_coords = self.geojson.coordinates
-        first_pass = True
-        last_inst = None
-
-        for inst in self.path:
-            if not inst.instruction_start_coordinate:
-                continue
-
-            end_coord = inst.instruction_start_coordinate.to_position()
-
-            try:
-                end_index = next(
-                    i
-                    for i, coord in enumerate(all_coords)
-                    if abs(coord[0] - end_coord[0]) < 1e-6 and abs(coord[1] - end_coord[1]) < 1e-6
-                )
-            except StopIteration:
-                continue
-
-            if first_pass:
-                first_pass = False
-                last_inst = inst
-                continue
-
-            yield (last_inst, LineString(coordinates=all_coords[: end_index + 1]))
-            all_coords = all_coords[end_index:]
-            last_inst = inst
-
-        if last_inst and len(all_coords) >= 2:
-            yield (last_inst, LineString(coordinates=all_coords))
-
-    def as_api_route_leg(self) -> RouteLeg:
-        default_location = (0, 0)  # TODO
-        insts = list(self.cut_linestring())
-
-        if self.mode:
-            mode = self.mode.to_mapbox()
-        else:
-            mode = TransportMode.tram
-
-        if not insts:
-            steps = [
-                RouteStep(
-                    maneuver=RouteManeuver(instruction="no instruction", location=default_location),
-                    duration=self.duration,
-                    distance=sum(inst.length for inst in self.path),
-                    geometry=self.geojson.dict(),
-                    properties={"mode": self.mode.value if self.mode else "tram"},
-                    mode=mode,
-                )
-            ]
-        else:
-            steps = [
-                RouteStep(
-                    maneuver=RouteManeuver(
-                        location=inst.instruction_start_coordinate.to_position(),
-                        instruction=inst.instruction or inst.name,
-                        modifier=inst.get_mapbox_modifier(),
-                    ),
-                    duration=inst.duration,
-                    distance=inst.length,
-                    geometry=geo,
-                    properties={"mode": self.mode.value if self.mode else "tram"},
-                    mode=mode,
-                )
-                for (inst, geo) in insts
-            ]
-
-        return RouteLeg(
-            duration=self.duration,
-            summary="todo",
-            steps=steps,
-            from_=TransportStop(location=self.geojson.coordinates[0]),
-            to=TransportStop(location=self.geojson.coordinates[-1]),
-        )
-
-
-class Journey(BaseModel):
-    arrival_date_time: datetime
-    departure_date_time: datetime
-    duration: int
-    distances: Distances
-    sections: List[Section]
-
-    @validator("arrival_date_time", "departure_date_time", pre=True)
-    def parse_date(cls, v):
-        return datetime.strptime(v, "%Y%m%dT%H%M%S")
-
-    def as_api_route(self) -> DirectionsRoute:
-        legs = [
-            sec.as_api_route_leg()
-            for sec in self.sections
-            if sec.sec_type != "waiting" and sec.sec_type != "tram"
-        ]
-
-        return DirectionsRoute(
-            duration=self.duration,
-            distance=self.distances.overall(),
-            start_time=datetime.isoformat(self.departure_date_time),
-            end_time=datetime.isoformat(self.arrival_date_time),
-            legs=legs,
-            geometry={
-                "type": "FeatureCollection",
-                "features": [
-                    {
-                        "type": "Feature",
-                        "properties": {
-                            "mode": leg.mode.value,
-                            **(
-                                {"lineColor": sec.display_informations.color}
-                                if sec.display_informations
-                                else {}
-                            ),
-                        },
-                        "geometry": {
-                            "type": "MultiLineString",
-                            "coordinates": [step.geometry["coordinates"] for step in leg.steps],
-                        },
-                    }
-                    for sec, leg in zip(self.sections, legs)
-                    for step in leg.steps
-                ]
-            },
-            summary=[
-                RouteSummaryPart(
-                    mode=TransportMode.bike,
-                    info=TransportInfo(
-                        num="42",
-                        direction="todo",
-                        lineColor="ff0000",
-                        network="RATP",
-                    ),
-                    distance=1337,
-                    duration=42,
-                ),
-                RouteSummaryPart(
-                    mode=TransportMode.walk,
-                    info=None,
-                    distance=1337,
-                    duration=42,
-                ),
-                RouteSummaryPart(
-                    mode=TransportMode.bike,
-                    info=TransportInfo(
-                        num="0",
-                        direction="todo",
-                        lineColor="00ff00",
-                        network="RATP",
-                    ),
-                    distance=1337,
-                    duration=42,
-                ),
-            ],
-        )
-
-
-class NavitiaResponse(BaseModel):
-    journeys: List[Journey] = []
-
-    def as_api_response(self) -> DirectionsResponse:
-        return DirectionsResponse(
-            status="ok",  # TODO,
-            data=DirectionsData(routes=[journey.as_api_route() for journey in self.journeys]),
-        )
 
 
 class MapboxAPIExtraParams(BaseModel):
